+++ pages/studio.jsx
@@ -1,13 +1,15 @@
 import { Fragment } from 'react';
-import Work from '../components/Work';
+import StudioFront from '../components/StudioFront';
 import PageContainer from '../components/PageContainer';
 import Contact from '../components/Contact';
+import Process from '../components/Process';
 
-export default function WorkPage() {
+export default function StudioPage() {
   return (
     <Fragment>
+      <StudioFront />
       <PageContainer>
-        <Work isTop />
+        <Process />
       </PageContainer>
       <PageContainer isDark>
         <Contact />
