+++ pages/team.jsx
@@ -1,6 +1,6 @@
-import Team from '../components/Team';
-import PageContainer from '../components/PageContainer';
 import Contact from '../components/Contact';
+import PageContainer from '../components/PageContainer';
+import Team from '../components/Team';
 
 export default function TeamPage() {
   return (
